feat(phase-3): unit test suites, CI jobs, and repo polish
Service unit tests (68 tests total across 4 suites):
- backend: 21 vitest + supertest tests (auth, health, modules, admin-logs)
- duckdb-service: 15 pytest tests (health, modules, nests, progress)
- image-service: 24 pytest tests (health, upload pipeline, module logs)
- homepage: 8 vitest + jsdom smoke tests (key React pages)

Bug fix:
- duckdb-service/models/progress.py: ClassificationOutput cell values
  typed as float (was int), matching the int(sealed * 100) route logic

CI (.github/workflows/tests.yml):
- Added 4 new parallel jobs (backend-unit, duckdb-unit, image-unit,
  homepage-unit), each capped at 5 min; all run alongside the existing
  esp-native, esp-firmware, e2e-pipeline jobs

Dev infrastructure:
- image-service/requirements-dev.txt: add -r requirements.txt so CI
  pip install actually installs Flask and friends
- .editorconfig, .prettierrc, .prettierignore added
- .github/ISSUE_TEMPLATE/ and PULL_REQUEST_TEMPLATE.md added

Documentation:
- ARCHITECTURE.md and CONTRIBUTING.md added (root-level entry points)
- documentation/architecture.md and api-usage.md refreshed

+++ duckdb-service/models/progress.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 class ClassificationOutput(BaseModel):
     modul_id: str
-    classification: Dict[str, Dict[int, int]]
+    classification: Dict[str, Dict[int, float]]
 
 
 BEE_TYPE_MAP = {
